tried to convert sharedAssets and kinda did but also very unsure what's going on with createReadStream..... so yeah
