+++ package.json
@@ -31,6 +31,7 @@
 		"email-validator": "^2.0.4",
 		"expo-server-sdk": "^3.6.0",
 		"express": "^4.18.1",
+		"fs-capacitor": "^8.0.0",
 		"global": "^4.4.0",
 		"graphql": "^15.3.0",
 		"graphql-type-json": "^0.3.2",
